generate CE-like plots with title and label
The 2D fingerprint plot now (again) contains a d_e and d_i as used to in
CrystalExplorer.  Gnuplot extracts the stem of the input file to provide
an automatic label on the plot, too.  The now explicit definition of the
point size in the scatter plot aims to prevent accidental overlap of the
'tiles' to lower the risk loosing visual information.