diff_finger.py, concept study
I'm curious if/how much efficiency of the computation is affected here
if the computation of the difference map relies on Python and numpy
instead of the C program.  Numpy isn't part of the standard library
shipped with Python, on the other hand offers a more efficient work
with matrices than standard Python.