reverse logic in array substraction
The previous definition to subtract the z values was not aligned with the
intended approach.  Inspection of diff*.dat files computed by diff_finger.c
and (previous) diff_finger.py with vimdiff revealed results of opposite
sign for the (previous) Python script vs. the compiled c executable by Rohl
and Raiteri.

This did not affect the computation of the ('Ruby') difference number,
however was equivalent to a mutual swap of de and di when visualizing
the difference map.