Windows moderator, start numpy free approach
This starts to remove the numpy dependency of the moderator script for
Windows.  On one hand, this aims to render the moderator more portable
because numpy is not member of Python's standard library.  On the other,
this step is going to be penalized by lower performance (though, a user
interested more in performance than portability likely continues to use
the compiled languages).