fingerprint_[Heron, Kahan, RR].py, concept study
There are (at least) three approaches how the surfaces triangulated by
CrystalExplorer eventually yield a 2D fingerprint.  So as a concept
study, the approach following Heron's formula, Kahan's approximiation,
and by Paolo Raiteri/Andrew Rohl (RR) is tested independent of the
other ones.  For now, the pace of computation is not as important than
the ease to use the script without (newly) introduced dependency to a
third party module; first objective is to probe how well the results
based on a 'Python only' match the ones using the Fortran approach by
Paolo Raiteri/Andrew Rohl.