fingerprint_[Heron, Kahan, RR].py, concept studies
Three Python scripts to probe the generation of 2D Hirshfeld
surface normalized fingerprints as an alternative to work
with Fortran.  Intentionally written only with modules of the
standard library for an increased portability, especially
with -- but not a priori limited to -- pypy in mind to obtain
an acceptable rate of processing the data.  The initial aim
to perform the core of the computational work in only one language
which is easy to access without dropping too much speed hence
seems to be reached.

This is a selective squash merge from branch fortran_py.