moderator scripts, backup map visualization
Shall gnuplot be absent, the scripts may relay the generation
of surveying image files to matplotlib (assisted by numpy).
Both are not part of the standard library, but do not present
a problem, e.g., to pypy without numpy as their import happenz
only if entering the corresponding branch of the script.

So far, the image generation with gnuplot remains preferred
over the present matplotlib setup.