moderator (Windows) hyphenated to Kahan approach
Alternative to the the Fortran / fingerprint.f90 based computation of 2D Hirshfeld
surface fingerprints, the Python-based analysis of the Hirshfeld surface with the
Kahan approach.  Different to other methods, the script containing the instructions
to calculate the triangle surfaces is kept separate (fingerprint_Kahan.py) from this
moderator script should testing the other approaches (Heron formula, or the orginal
algorithm used by Andrew Rohl and Paolo Raiteri) be of interest for comparison.