Python-externalized normalization
The computation of normalized 2D Hirshfeld surface fingerprints may be relayed to either Fortran
and fingerprint.f90 (requiring gcc or gfortran), or to scripts written in Python allowing to
continue the computation in the Python-unified approach.  Preferred -- thus encoded in the
moderator scripts -- is the computation of the triangle surfaces by the Kahan approach; but both
one script echoing the approach by Rohl and Raiteri as well as one based on Heron formula are
included.  The fingerprint_[heron, kahan, rr].py scripts equally may be used independently from
the  moderator scripts, too.