fingerprint scripts in classes
This commit reorganizes the functions of the three fingerscript in classes
to allow use of variables and constants as 'global'.

The fingerprint scripts now use, e.g. fingerprint_kahan.main() to start
processing the .cxs data file(s) by identification up and including the
deposit of the corresponding fingerprint .dat file. The two moderator
scripts are updated accordingly (though in commits separate to this one).