hirshfeld_moderator.py, bug-fix gcc in Windows
The compilation of diff_finger.c in Windows with gcc yields an executable
with the file extension of ".exe" -- even if this was not explicitly
stated in the Python script.  Of course, this prevented its transfer to
the data, its work to compute the difference maps, and eventual space
cleaning.

As a rebase of branch gcc_fix in branch gcc_fix_rebase, this issue is
resolved.