hirshfeld_moderator.py, rm shebang
The heading shebang line was removed.  For one, a deployment
in Windows will not consider this entry.  For two, according
to "Automate the boring stuff with Python" by Al Sweigart,
the shebang in MacOS might differ from the one in Linux.  For
three, the earlier form was reaching out for python3 -- which
may contradict the earlier stated aim to render the script as
accessible to Python (such as CPython3 _or_  CPython2, _or_
pypy).  As a cosmetic issue, absence of this form of shebang
trggers Xubuntu to display the Python symbol on top of the
file, too.