moderator, bugfix for work in Windows
The same compilation command issued to gcc about diff_finger.c
yields executables of different names in Linux, and Windows.
In Windows, not accepting an extension less file, the compilation
yields an executable with explicit -- and added -- file extension
'.exe'.  In Windows, this previously hindered the executable to
be copied to the data, and to compute the difference maps.

Assuming a specific shebang to indicate Python 3 is too strict
when aiming for a moderator to be equally used in Python 2, the
shebang was dropped.  (Xubuntu's file manager Thunar anyway is
able to recognize this as Python source code and displays the
file with the intertwinned snakes -- which it doesn't yet(?) for
files including the python3 specific shebang.)