hirshfeld_moderator, bug fix, gnuplot plot windows
A rebase merge from gnuplot_fix_rebased, based on gnuplot_fix, to enable
gnuplot's action -- in Windows -- to provide the surveying .png of small
scale, as well as the higher quality bitmap .png and vector .pdf with
their bells and whistles.

This brings the two OSes to a similar level, not only about the core
computations to generate the normalized 2D Hirshfeld surface fingerprint
maps, the difference maps, and the difference number _computed_; now,
both Linux and Windows are capable to reach out for either the still
preferred gnuplot, or alternative Python's matplotlib to visualize the
fingerprint maps and difference maps.