explicit format definition, (x,y) in difference map
The previous definition reported the (x,y) coordinates of the
difference map with a large number of decimals.  By inspection of
CrystalExplorer's fingerprint maps and intermediate files of this
project -- already at highest resolution -- one recognizes the
increment along x and y is by 0.01 A, but not 0.000001 A.  To
round these coordinates from e.g.,

```
0.850000 1.170000  0.000000
0.850000 1.180000 -0.001398
0.850000 1.190000  0.013031
0.850000 1.200000  0.011165
0.850000 1.210000 -0.003442
0.850000 1.220000  0.000000
```

to only two decimals, i.e.,

```
0.85 1.17  0.000000
0.85 1.18 -0.001398
0.85 1.19  0.013031
0.85 1.20  0.011165
0.85 1.21 -0.003442
0.85 1.22  0.000000
```

equally reduces the file size by about 28% without obstruction
to plot a visualization, nor to compute the difference number.