add a difference number implementation in AWK
With focus on processing one file at a time, the computation of
the difference number was implemented in AWK to probe if/how much
an implementation in language different to Ruby (used by Rohl and
Raiteri) could be faster.  While the source code is short, a future
proof of concept likely is going to focus on Fortran.