use an explicit fixed format to report the 2D maps
In previous commits, the normalized 2D maps by fingerprint.f90 did
not use an explicit format of fixed precission to report de, di,
and distance(idi,ide).  This lead to leading/trailing spaces and
in use of more decimals than the 0.01 A increments require.  In the
example of BZAMID01.cxs, for example

```
  0.79000000000000004       0.97999999999999998        0.0000000000000000
  0.79000000000000004       0.98999999999999999        0.0000000000000000
  0.79000000000000004        1.0000000000000000        0.0000000000000000
  0.79000000000000004        1.0100000000000000        0.0000000000000000
```

and

```
  0.79000000000000004        1.1299999999999999        0.0000000000000000
  0.79000000000000004        1.1400000000000001        1.0291002179768300E-002
  0.79000000000000004        1.1499999999999999        0.0000000000000000
```

With the now explicit format definition of fixed precision, these
unwanted variations do no longer occur:

```
 0.79 0.98 0.000000000000
 0.79 0.99 0.000000000000
 0.79 1.00 0.000000000000
 0.79 1.01 0.000000000000
```

 and

 ```
 0.79 1.13 0.000000000000
 0.79 1.14 0.010291002180
 0.79 1.15 0.000000000000
```

The introduction of the fixed format aims to ease the subsequent
computations while offering enough decimals for Fortran's double
precision format.  As checked with BZAMID01 and BZAMID02 when
requesting the standard 2D maps, the difference map computed with
the normalized maps in standard map range of this commit does not
differ from the one computed by fingerprint.f90 prior to this very
commit.  In addition, the file size of the normalized 2D map now
is less than half of the normalized 2D map reported earlier when
using the old format.