mute a line separtor
Previous commits generated a blank separator line when 'sweeping
the next line' in the 2D representation:

```
 0.40 2.59 0.000000000000
 0.40 2.60 0.000000000000

 0.41 0.40 0.000000000000
 0.41 0.41 0.000000000000
```

Checks with gnuplot reveal that this not necessary, and a report
like

```
 0.40 2.59 0.000000000000
 0.40 2.60 0.000000000000
 0.41 0.40 0.000000000000
 0.41 0.41 0.000000000000
```

works equally well.  While savings in file size by this change
are minute, this representation likely is easier to process as
an array (e.g., in Fortran), than the earlier format.