adjust output format to C program
The previously used format definition of the difference map was a
different one then the one used by the C program.  Now, both use
the same format.  In a test run with the benzamid reference data
BZAMID01.dat and BZAMID11.dat (extended map range 0.40 to 3.00 A),
the difference map provided by both (68121 lines) differed five
times in the intensity value / z-column at the sixth decimal, and
once for the (3.00, 3.00) x,y coordinate (the very last entry).
These differences are deemed insignificant for the subsequent
computation of the difference number.