test an array-based diffnumber computation
At time of compilation, the maximal size of a difference map -
extended range, di and de between 0.40 and 3.00 A in increments of
0.01 A - is known.  So then an array of zeroes can be preallocated
and updated by the actual data of the map which to prepare an
array based computation of the difference number.  This approach
can be faster than a linewise read and recomputation of the number
by `sum_abs_diff.f90`.