prevent unwanted potential data type conversion
The computation of bin indices idi and ide trades with real data,
however subsequent decisions anticipate a return of integers.  As
checked by diff -s on BZAMID01.dat as an extended map for reference
data set BZAMID01.cxs, the modifification in the source code does
not affect the output written.