refactor(diff_finger.py): reorganize file read
Raw data are read directly into a numpy array.  Prior to the
computation of a difference map there is a pairwise check of
the arrays to run the computation of a difference map only
with data deemed suitable by lower / upper limit of d_e and
line count of the arrays.

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>