perf(diff_finger.py): reduce read frequency ref_file
For the computation of difference maps AB, AC, AD; BC, BD; CD -
the first file deemed to be a reference file of the two only now
is read once instead thrice (AB, AC, AD), or twice (BC, BD).
Equally, an additional check is installed to prevent ValueErrors
while reading the raw data to stop the execution of the script.

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>